requirements.txt
chromadb
tiktoken
langchain
langchain-community
langchain-openai
langchain_anthropic
langchain-nomic
sentence_transformers
google-api-python-client
requests
openai
yt_dlp
langchain-chroma
tqdm
unstructured
python-fasthtml
uvicorn==0.30.1